Reorganization and Validation
